+++ compiler/daml-lf-proto/BUILD.bazel
@@ -4,28 +4,106 @@
 load("//bazel_tools:haskell.bzl", "da_haskell_library")
 
 da_haskell_library(
-    name = "daml-lf-proto",
-    srcs = glob(["src/**/*.hs"]),
+    name = "daml-lf-util",
+    srcs =
+        [
+            "src/DA/Daml/LF/Mangling.hs",
+            "src/DA/Daml/LF/Proto3/Error.hs",
+            "src/DA/Daml/LF/Proto3/Util.hs",
+        ],
     hackage_deps = [
         "base",
         "bytestring",
-        "containers",
-        "cryptonite",
         "either",
-        "lens",
-        "memory",
-        "mtl",
-        "proto3-suite",
-        "scientific",
-        "template-haskell",
         "text",
-        "transformers",
-        "unordered-containers",
-        "vector",
     ],
-    src_strip_prefix = "src",
     visibility = ["//visibility:public"],
     deps = [
+        "//compiler/daml-lf-ast",
+        "//daml-lf/archive:daml_lf_dev_archive_haskell_proto",
+    ],
+)
+
+da_haskell_library(
+    name = "daml-lf-proto-decode",
+    srcs = [
+        "src/DA/Daml/LF/Proto3/Archive/Decode.hs",
+        "src/DA/Daml/LF/Proto3/Decode.hs",
+        "src/DA/Daml/LF/Proto3/DecodeV1.hs",
+    ],
+    hackage_deps = [
+        "base",
+        "containers",
+        "lens",
+        "cryptonite",
+        "memory",
+        "bytestring",
+        "mtl",
+        "proto3-suite",
+        "text",
+        "vector",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":daml-lf-util",
+        "//compiler/daml-lf-ast",
+        "//daml-lf/archive:daml_lf_dev_archive_haskell_proto",
+        "//libs-haskell/da-hs-base",
+    ],
+)
+
+da_haskell_library(
+    name = "daml-lf-proto-encode",
+    srcs = [
+        "src/DA/Daml/LF/Proto3/Archive/Encode.hs",
+        "src/DA/Daml/LF/Proto3/Encode.hs",
+        "src/DA/Daml/LF/Proto3/EncodeV1.hs",
+    ],
+    hackage_deps = [
+        "base",
+        "bytestring",
+        "cryptonite",
+        "memory",
+        "containers",
+        "mtl",
+        "lens",
+        "text",
+        "vector",
+        "proto3-suite",
+        "unordered-containers",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":daml-lf-util",
+        "//compiler/daml-lf-ast",
+        "//daml-lf/archive:daml_lf_dev_archive_haskell_proto",
+        "//libs-haskell/da-hs-base",
+    ],
+)
+
+da_haskell_library(
+    name = "daml-lf-proto",
+    srcs = [
+        "src/DA/Daml/LF/Proto3/Archive.hs",
+    ],
+    hackage_deps = [
+        "base",
+        "containers",
+        "cryptonite",
+        "memory",
+        "bytestring",
+        "mtl",
+        "lens",
+        "text",
+        "vector",
+        "proto3-suite",
+        "unordered-containers",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":daml-lf-proto-decode",
+        ":daml-lf-proto-encode",
+        ":daml-lf-util",
         "//compiler/daml-lf-ast",
         "//daml-lf/archive:daml_lf_dev_archive_haskell_proto",
         "//libs-haskell/da-hs-base",
